+++ CHANGELOG.md
@@ -8,9 +8,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 - linear source inversion *lsrcinv* in *nessi.signal*
+- *normalize* method for SUdata()
 
 ### Modified
-- *pfilter* in *SUdata()* class
+- *pfilter* method for SUdata()
 - *sin2filter* function in *filtering.py*
 
 ## [0.2.2] - 2018-12-05